routes, images, utils.js  local and package import separations and renaming applied as suggested
