+++ src/App.jsx
@@ -1,26 +1,15 @@
 import './App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Navbar from './components/Navbar';
 import Home from './pages/Home';
-import About from './pages/About';
-import Contact from './pages/Contact';
-import Login from './pages/Login';
-import SignUp from './pages/SignUp';
-import NotFound from './pages/NotFound';
 import Footer from './components/Footer';
+import { HOME_ROUTE } from './routes';
 
 function App() {
   return (
-    <div className=" ">
+    <div>
       <Router>
-        <Navbar />
         <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="about" element={<About />} />
-          <Route path="contact" element={<Contact />} />
-          <Route path="login" element={<Login />} />
-          <Route path="signup" element={<SignUp />} />
-          <Route path="*" element={<NotFound />} />
+          <Route path={HOME_ROUTE} element={<Home />} />
         </Routes>
         <Footer />
       </Router>
