--- src/pages/About.js
@@ -1,9 +1,0 @@
-function About() {
-  return (
-    <div className="h-64 container mx-auto px-8 text-3xl text-center flex items-center justify-center">
-      About
-    </div>
-  );
-}
-
-export default About;
