--- src/pages/Contact.js
@@ -1,9 +1,0 @@
-function Contact() {
-  return (
-    <div className="h-64 container mx-auto px-8 text-3xl text-center flex items-center justify-center">
-      Contact
-    </div>
-  );
-}
-
-export default Contact;
