--- src/pages/Login.js
@@ -1,9 +1,0 @@
-function Login() {
-  return (
-    <div className="h-64 container mx-auto px-8 text-3xl text-center flex items-center justify-center">
-      Login
-    </div>
-  );
-}
-
-export default Login;
