--- src/pages/SignUp.js
@@ -1,9 +1,0 @@
-function SignUp() {
-  return (
-    <div className="h-64 container mx-auto px-8 text-3xl text-center flex items-center justify-center">
-      SignUp
-    </div>
-  );
-}
-
-export default SignUp;
