+++ src/routes.js
@@ -1,2 +1,7 @@
-export const HOME_ROUTE = 'home';
+export const HOME_ROUTE = '/';
 export const CONTACT_ROUTE = 'contact';
+export const ABOUT_ROUTE = 'about';
+export const LOGIN_ROUTE = 'login';
+export const SIGN_UP_ROUTE = 'signup';
+export const PROJECT_ROUTE = 'project';
+export const NOT_FOUND_ROUTE = '*';
